Added interpreter creation/simulation
The new classes can handle the creation and simulation of an interpreter
within a workspace. Using bison and flex one can generate the files that
describe how an interpreter works.
These files can be compiled into an interpreter using GCC.
Finally, an interpreter can be simulated using an input file from the
workspace.
In order to perform all these operations, the workspace and its contents
are copied to a folder with the same name in the system's temporary
storage. Once the operation is done, the results are copied back to the
application's storage.
Added necessary parameters and registered the new UtilityHandler service