Changed the way custom listeners work
Instead of having the listeners check for routes (which had started
getting messy and depended on the environment), checking whether the
controller is an instance of a specific interface is more
straightforward. Each controller can implement more than one interfaces,
so more than one checks can be applied to the corresponding requests.
FilterControllerEvents though cannot have their response modified. As a
result, the listeners now throw exceptions which are handled by a
separate listener.

- Added interfaces that allow better handling of different controllers.
- Modified the listeners, so that all of them listen to controller
  events, according to the approach described above.
- Added an exception listener, which enables modification of an event's
  response on demand.
- Updated services registry.