Allow admins to delete teams (#5)
* Added initial action code

* Moved sanity check to controller

Move the last-team-member check from the team manager to the controller.
This allows the team manager method to be used by both simple users and
admins.

* Added UserManager class

The class is meant for user related operations. Currently, only multiple
user deletion is supported.

* Added the rest of the code for team deletion action

* Fixed typos

* Added missing flush operation for user deletion

* Added frontend for team deletion

A 'Delete team' button is now available in every search result. The
button is also connected to the javascript code that handles the team
deletion request.

* Added missing action name suffix

* Added a confirmation modal for team deletion

* Confirmation modal visual improvement